pyproject.toml
[project]
name = "next-plan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "gradio>=6.7.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-sarvam-integration>=0.1.7",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.9",
    "pydantic==2.7.4",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
